pyproject.toml
[project]
name = "call-center"
version = "0.1.0"
description = "Call center quality grading system using multi-agent AI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "openai>=1.40.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "streamlit>=1.35.0",
    "mutagen>=1.47.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "deepgram-sdk>=5.0.0",
    "guardrails-ai>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.setuptools]
packages = ["app", "agents", "graph", "schemas", "services", "utils"]
